fix: Improved validation and better structure
