+++ tools/market-config-cli/src/config/validator.rs
@@ -265,20 +265,34 @@ pub enum TokenMetadata {
     Nep245(MultiTokenMetadata),
 }
 
-async fn fetch_metadata<T: templar_common::asset::AssetClass>(
+#[derive(Debug)]
+pub(crate) enum ResolvedAssetParts {
+    Nep141(AccountId),
+    Nep245 {
+        contract_id: AccountId,
+        token_id: String,
+    },
+}
+
+/// # Errors
+pub async fn fetch_metadata<T: templar_common::asset::AssetClass>(
     client: &JsonRpcClient,
     asset: &FungibleAsset<T>,
 ) -> CliResult<TokenMetadata> {
-    if let Some((contract_id, token_id)) = asset.clone().into_nep245() {
-        let (resolved_contract, resolved_token_id) =
-            resolve_nep141_or_nep245_parts(contract_id, &token_id)?;
-        let token_id = resolved_token_id.ok_or_else(|| {
-            CliError::Validation(format!("Missing NEP-245 token id for asset '{asset}'"))
-        })?;
-        // return multitoken_metadata(client, resolved_contract, token_id).await;
-
-        let metadata = multitoken_metadata(client, resolved_contract, token_id).await?;
-        return Ok(TokenMetadata::Nep245(metadata));
+    if let Some((_, token_id)) = asset.clone().into_nep245() {
+        match resolve_nep141_or_nep245_parts(&token_id)? {
+            ResolvedAssetParts::Nep141(account_id) => {
+                let metadata = ft_metadata(client, account_id).await?;
+                return Ok(TokenMetadata::Nep141(metadata));
+            }
+            ResolvedAssetParts::Nep245 {
+                contract_id,
+                token_id,
+            } => {
+                let metadata = multitoken_metadata(client, contract_id, token_id).await?;
+                return Ok(TokenMetadata::Nep245(metadata));
+            }
+        }
     }
 
     let id: AccountId = asset.contract_id().as_ref().parse().map_err(|e| {
@@ -297,33 +311,43 @@ pub async fn check_asset_existence<T: templar_common::asset::AssetClass>(
     client: &JsonRpcClient,
     asset: &FungibleAsset<T>,
 ) -> CliResult {
-    if let Some((contract_id, token_id)) = asset.clone().into_nep245() {
-        let (resolved_contract, resolved_token_id) =
-            resolve_nep141_or_nep245_parts(contract_id, &token_id)?;
-        // NEP-245 multi-token
-        view_account(client, resolved_contract.clone())
-            .await
-            .map_err(|e| {
-                CliError::Validation(format!(
-                    "Account check failed for '{resolved_contract}': {e}"
-                ))
-            })?;
+    if let Some((_, token_id)) = asset.clone().into_nep245() {
+        match resolve_nep141_or_nep245_parts(&token_id)? {
+            ResolvedAssetParts::Nep141(account_id) => {
+                view_account(client, account_id.clone())
+                    .await
+                    .map_err(|e| {
+                        CliError::Validation(format!(
+                            "Account check failed for '{account_id}': {e}"
+                        ))
+                    })?;
+            }
+            ResolvedAssetParts::Nep245 {
+                contract_id,
+                token_id,
+            } => {
+                // NEP-245 multi-token
+                view_account(client, contract_id.clone())
+                    .await
+                    .map_err(|e| {
+                        CliError::Validation(format!(
+                            "Account check failed for '{contract_id}': {e}"
+                        ))
+                    })?;
 
-        if let Some(resolved_token_id) = resolved_token_id {
-            let metadata =
-                multitoken_metadata(client, resolved_contract, resolved_token_id.clone())
+                let metadata = multitoken_metadata(client, contract_id, token_id.clone())
                     .await
                     .map_err(|e| {
                         CliError::Validation(format!("Failed to fetch NEP-245 metadata: {e}"))
                     })?;
-            let metadata_id = metadata
-                .id
-                .as_deref()
-                .ok_or_else(|| CliError::Validation("NEP-245 metadata missing id field".into()))?;
-            if metadata_id != resolved_token_id {
-                return Err(CliError::Validation(format!(
-                    "NEP-245 token id mismatch: expected '{resolved_token_id}', got '{metadata_id}'"
-                )));
+                let metadata_id = metadata.id.as_deref().ok_or_else(|| {
+                    CliError::Validation("NEP-245 metadata missing id field".into())
+                })?;
+                if metadata_id != token_id {
+                    return Err(CliError::Validation(format!(
+                        "NEP-245 token id mismatch: expected '{token_id}', got '{metadata_id}'"
+                    )));
+                }
             }
         }
     }
@@ -343,10 +367,7 @@ pub async fn check_asset_existence<T: templar_common::asset::AssetClass>(
     Ok(())
 }
 
-pub(crate) fn resolve_nep141_or_nep245_parts(
-    contract_id: AccountId,
-    token_id: &str,
-) -> CliResult<(AccountId, Option<String>)> {
+pub(crate) fn resolve_nep141_or_nep245_parts(token_id: &str) -> CliResult<ResolvedAssetParts> {
     if let Some(stripped) = token_id.strip_prefix("nep141:") {
         let mut parts = stripped.splitn(2, ':');
         let Some(ft_contract) = parts.next() else {
@@ -361,7 +382,7 @@ pub(crate) fn resolve_nep141_or_nep245_parts(
             ))
         })?;
 
-        return Ok((ft_contract, None));
+        return Ok(ResolvedAssetParts::Nep141(ft_contract));
     }
 
     if let Some(stripped) = token_id.strip_prefix("nep245:") {
@@ -382,10 +403,15 @@ pub(crate) fn resolve_nep141_or_nep245_parts(
                 "Invalid NEP-245 contract id in token '{token_id}': {e}"
             ))
         })?;
-        return Ok((mt_contract, Some(mt_token_id.to_string())));
+        return Ok(ResolvedAssetParts::Nep245 {
+            contract_id: mt_contract,
+            token_id: mt_token_id.to_string(),
+        });
     }
 
-    Ok((contract_id, Some(token_id.to_string())))
+    Err(CliError::Validation(format!(
+        "Token id '{token_id}' must be prefixed with 'nep141:' or 'nep245:'"
+    )))
 }
 
 /// Validate that decimals are within reasonable bounds (typically 0-24 for NEAR)
@@ -522,14 +548,16 @@ mod tests {
 
     #[test]
     fn resolve_nep141_or_nep245_parts_supports_nep141_prefix() {
-        let contract_id = AccountId::from_str("intents.near").unwrap();
         let token_id = "nep141:usdc.near:anything".to_string();
 
-        let (resolved_contract, resolved_token_id) =
-            resolve_nep141_or_nep245_parts(contract_id, &token_id).unwrap();
+        let resolved = resolve_nep141_or_nep245_parts(&token_id).unwrap();
 
-        assert_eq!(resolved_contract, AccountId::from_str("usdc.near").unwrap());
-        assert!(resolved_token_id.is_none());
+        match resolved {
+            ResolvedAssetParts::Nep141(account_id) => {
+                assert_eq!(account_id, AccountId::from_str("usdc.near").unwrap());
+            }
+            ResolvedAssetParts::Nep245 { .. } => panic!("expected nep141 resolution"),
+        }
     }
 
     #[test]
@@ -537,11 +565,18 @@ mod tests {
         let contract_id = AccountId::from_str("intents.near").unwrap();
         let token_id = "nep245:intents.near:foo".to_string();
 
-        let (resolved_contract, resolved_token_id) =
-            resolve_nep141_or_nep245_parts(contract_id.clone(), &token_id).unwrap();
+        let resolved = resolve_nep141_or_nep245_parts(&token_id).unwrap();
 
-        assert_eq!(resolved_contract, contract_id);
-        assert_eq!(resolved_token_id, Some("foo".to_string()));
+        match resolved {
+            ResolvedAssetParts::Nep245 {
+                contract_id: resolved_contract,
+                token_id: resolved_token_id,
+            } => {
+                assert_eq!(resolved_contract, contract_id);
+                assert_eq!(resolved_token_id, "foo".to_string());
+            }
+            ResolvedAssetParts::Nep141(_) => panic!("expected nep245 resolution"),
+        }
     }
 
     #[test]
@@ -549,22 +584,27 @@ mod tests {
         let contract_id = AccountId::from_str("intents.near").unwrap();
         let token_id = "nep245:intents.near:bar".to_string();
 
-        let (resolved_contract, resolved_token_id) =
-            resolve_nep141_or_nep245_parts(contract_id.clone(), &token_id).unwrap();
+        let resolved = resolve_nep141_or_nep245_parts(&token_id).unwrap();
 
-        assert_eq!(resolved_contract, contract_id);
-        assert_eq!(resolved_token_id, Some("bar".to_string()));
+        match resolved {
+            ResolvedAssetParts::Nep245 {
+                contract_id: resolved_contract,
+                token_id: resolved_token_id,
+            } => {
+                assert_eq!(resolved_contract, contract_id);
+                assert_eq!(resolved_token_id, "bar".to_string());
+            }
+            ResolvedAssetParts::Nep141(_) => panic!("expected nep245 resolution"),
+        }
     }
 
     #[test]
-    fn resolve_nep141_or_nep245_parts_passthrough_when_no_prefix() {
-        let contract_id = AccountId::from_str("intents.near").unwrap();
+    fn resolve_nep141_or_nep245_parts_rejects_missing_prefix() {
         let token_id = "plain-token".to_string();
 
-        let (resolved_contract, resolved_token_id) =
-            resolve_nep141_or_nep245_parts(contract_id.clone(), &token_id).unwrap();
-
-        assert_eq!(resolved_contract, contract_id);
-        assert_eq!(resolved_token_id, Some(token_id));
+        let err = resolve_nep141_or_nep245_parts(&token_id).unwrap_err();
+        assert!(err
+            .to_string()
+            .contains("must be prefixed with 'nep141:' or 'nep245:'"));
     }
 }
