--- tools/market-config-cli/src/editor/mod.rs
@@ -1,5 +1,0 @@
-pub use crate::common::prompt::market_prompter::MarketPrompter as ConfigEditor;
-
-pub mod utils {
-    pub use crate::common::prompt::utils::*;
-}
