--- tools/market-config-cli/src/interactive/mod.rs
@@ -1,6 +1,0 @@
-pub use crate::common::prompt::market_prompter::MarketPrompter as InteractivePrompt;
-
-pub mod prompts {
-    pub use crate::common::prompt::market_prompter::MarketPrompter as InteractivePrompt;
-    pub use crate::common::prompt::utils::parse_price_id;
-}
