+++ tools/market-config-cli/src/ui/mod.rs
@@ -1,2 +1,1 @@
 pub mod prompt;
-pub mod shared;
