+++ tools/market-config-cli/src/ui/prompt/types.rs
@@ -1,77 +1,7 @@
-use crate::{interactive::prompts::parse_price_id, CliError, CliResult};
-use dialoguer::{theme::ColorfulTheme, Input};
-use near_sdk::{json_types::U128, AccountId};
+use crate::{CliError, CliResult};
 use serde_json::Value;
-use std::{collections::HashMap, fmt, str::FromStr};
-use templar_common::{
-    asset::{AssetClass, FungibleAsset},
-    fee::Fee,
-    interest_rate_strategy::InterestRateStrategy,
-    number::Decimal,
-    oracle::pyth::PriceIdentifier,
-};
-
-/// # Errors
-pub fn parse_asset_input<T: AssetClass>(value: &str, field: &str) -> CliResult<FungibleAsset<T>> {
-    match value.parse::<FungibleAsset<T>>() {
-        Ok(asset) => Ok(asset),
-        Err(_) => AccountId::from_str(value)
-            .map(FungibleAsset::nep141)
-            .map_err(|e| CliError::InvalidInput(format!("Invalid {field}: {e}"))),
-    }
-}
-
-/// # Errors
-pub fn price_id_from_input(value: &str) -> CliResult<PriceIdentifier> {
-    parse_price_id(value.trim())
-        .map_err(|e| CliError::InvalidInput(format!("Invalid price ID '{value}': {e}",)))
-}
-
-/// # Errors
-pub fn prompt_decimal(
-    theme: &ColorfulTheme,
-    prompt: &str,
-    default: &str,
-    field: &str,
-) -> CliResult<Decimal> {
-    let value: String = Input::with_theme(theme)
-        .with_prompt(prompt)
-        .default(default.to_string())
-        .interact_text()
-        .map_err(std::io::Error::other)?;
-    Decimal::from_str(&value)
-        .map_err(|_| CliError::InvalidInput(format!("Invalid decimal for {field}: {value}")))
-}
-
-/// Prompt for an integer decimal count with inline bounds checking (0-24).
-/// # Errors
-pub fn prompt_decimals(
-    theme: &ColorfulTheme,
-    prompt: &str,
-    default: i32,
-    field: &str,
-) -> CliResult<i32> {
-    let value: i32 = Input::with_theme(theme)
-        .with_prompt(prompt)
-        .default(default)
-        .interact_text()
-        .map_err(std::io::Error::other)?;
-
-    if (0..=24).contains(&value) {
-        Ok(value)
-    } else {
-        Err(CliError::InvalidInput(format!(
-            "{field} must be between 0 and 24"
-        )))
-    }
-}
-
-pub fn fee_defaults<T: AssetClass>(fee: &Fee<T>) -> (usize, String) {
-    match fee {
-        Fee::Flat(amount) => (0, U128::from(*amount).0.to_string()),
-        Fee::Proportional(pct) => (1, pct.to_string()),
-    }
-}
+use std::{collections::HashMap, fmt};
+use templar_common::interest_rate_strategy::InterestRateStrategy;
 
 #[derive(Clone, Debug)]
 pub struct StrategyDefaults {
